integer overflow
# message auto-generated for no-merge-commit merge:
merge integer_overflow into br_develop_boostio

totalLen累加数据源为对端返回数据，没有做任何校验，存在整数溢出风险


Match-id-671f2554ec133b2ebbdd110cf2b08645bb4112fa

+++ src/interceptor/client/interceptor_fileio.cpp
@@ -272,7 +272,7 @@ ssize_t ProxyOperations::PwriteLargeInner(int fd, const void *buf, size_t count,
         ", location1:" << resp.address.loc.location[1] << ", address0 size:" << resp.address.address[0].size <<
         ", address1 size:" << resp.address.address[1].size << ", address num:" << resp.address.addressNum << ".");
 
-    uint32_t totalLen = 0;
+    uint64_t totalLen = 0;
     for (uint32_t i = 0; i < resp.address.addressNum; i++) {
         totalLen += resp.address.address[i].size;
     }
